style: banner 분리, emotion 태그 간소화
