+++ src/pages/Home/MainPetitions/styles.ts
@@ -1,0 +1,18 @@
+import styled from '@emotion/styled'
+
+const PetitionsWrapper = styled.div`
+  padding-bottom: 60px;
+  .petitions_title {
+    display: flex;
+    text-align: center;
+    padding-top: 4rem;
+    padding-bottom: 2rem;
+    border-bottom: 2px solid #ccc;
+    > span {
+      font-size: 20px;
+      font-weight: bold;
+    }
+  }
+`
+
+export { PetitionsWrapper }
